# api/openapi.yaml
openapi: 3.1.0
info:
  title: (S)AGE API
  version: 11.15.0
  description: |-
    (Sovereign) Agent Governed Experience — network REST API for memory
    submission, validation, and query.

    This machine-readable contract intentionally excludes the same-machine
    human CEREBRUM control plane under `/v1/dashboard/**`. Those routes are
    loopback-only management APIs, not remotely callable network APIs. Their
    authoritative contract is the “App-v23 CEREBRUM access-control routes”
    section of `docs/reference/rest-api.md`.
  contact:
    name: Dhillon Andrew Kannabhiran
    email: dhillon@levelupctf.com
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development

# ---------------------------------------------------------------------------
# Security — Ed25519 signature authentication via headers
# ---------------------------------------------------------------------------
security:
  - ed25519Auth: []
components:
...
  securitySchemes:
    ed25519Auth:
      type: apiKey
      in: header
      name: X-Agent-ID
      description: >-
        Ed25519 authentication is conveyed through three headers sent together:
        **X-Agent-ID** (hex-encoded public key), **X-Signature** (Ed25519
        signature of the canonical request payload), and **X-Timestamp** (unix
        epoch seconds). All three must be present on authenticated endpoints.
        Current clients should also send **X-Nonce**. After app-v17 activation,
        consensus binds delegated REST proofs to the exact canonical request,
        enforces the timestamp against block time, and consumes each proof once.

  # -----------------------------------------------------------------------
  # Parameters
  # -----------------------------------------------------------------------
...
  schemas:
...
    AccessRequestBody:
      type: object
      required:
        - target_domain
      properties:
        target_domain:
          type: string
        justification:
          type: string
        requested_level:
          type: integer
          description: 1=read, 2=read+write. Defaults to 1.
    AccessGrantRequest:
      type: object
      required:
        - grantee_id
        - domain
      properties:
        grantee_id:
          type: string
          description: Hex agent ID of the grantee.
        domain:
          type: string
        level:
          type: integer
          description: 1=read, 2=read+write. Defaults to 1.
        expires_at:
          type: integer
          format: int64
          description: Unix timestamp; 0=permanent.
        request_id:
          type: string
          description: Links to originating access request.
    AccessRevokeRequest:
      type: object
      required:
        - grantee_id
        - domain
      properties:
        grantee_id:
          type: string
        domain:
          type: string
        reason:
          type: string

    # -- Domains ---------------------------------------------------------------
    DomainRegisterRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
        parent:
          type: string
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs.
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type.
        title:
          type: string
          description: Short, human-readable summary.
        status:
          type: integer
          description: HTTP status code.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri
          description: URI reference identifying the specific occurrence.
        reason_code:
          type: string
          description: Stable sanitized machine-readable denial reason, when available.
        remedy:
          type: string
          description: Exact operator action that can make a later request valid.
        retryable:
          type: boolean
          description: Whether automatically repeating the same request is safe.
...
  responses:
    BadRequest:
      description: Invalid request body or parameters.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Unauthorized:
      description: Missing or invalid authentication headers.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Forbidden:
      description: >-
        Caller authenticated but lacks authorisation, or the request was
        rejected during on-chain finalisation (e.g. quorum-gated tx without
        an executed proposal). The `detail` field carries the upstream
        rejection reason.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    InternalError:
      description: Internal server error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
paths:
...
  /v1/access/request:
    post:
      operationId: accessRequest
      summary: Request read or read+write access to a domain.
      tags:
        - Access
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessRequestBody"
      responses:
        "201":
          description: Access request submitted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: pending
                  tx_hash:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/access/grant:
    post:
      operationId: accessGrant
      summary: Grant domain access. Caller must own the domain or be admin.
      tags:
        - Access
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessGrantRequest"
      responses:
        "201":
          description: Access granted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: granted
                  tx_hash:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/access/revoke:
    post:
      operationId: accessRevoke
      summary: Revoke domain access.
      tags:
        - Access
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessRevokeRequest"
      responses:
        "200":
          description: Access revoked.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: revoked
                  tx_hash:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/domain/register:
    post:
      operationId: domainRegister
      summary: Register a domain. Caller becomes owner.
      tags:
        - Domain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DomainRegisterRequest"
      responses:
        "201":
          description: Domain registered.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: registered
                  tx_hash:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/domain/{name}:
    get:
      operationId: getDomain
      summary: Get domain metadata.
      description: >-
        Ownership served from BadgerDB (chain-authoritative); description/created_at
        enriched from off-chain mirror.
      tags:
        - Domain
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Dotted-path domain name.
      responses:
        "200":
          description: Domain metadata.
          content:
            application/json:
              schema:
                type: object
                properties:
                  domain_name:
                    type: string
                  owner_agent_id:
                    type: string
                  parent_domain:
                    type: string
                  created_height:
                    type: integer
                    format: int64
                  description:
                    type: string
                  created_at:
                    type: string
                    format: date-time
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...